+++ samples/blocks/pinball/Pinball.cs
@@ -16,7 +16,7 @@ using System.Numerics;
 
 using Blitter;
 using Blitter.Bits;
-using Blitter.Blocks;
+using Blitter.Blocks2D;
 
 using SkiaSharp;